Complete rewrote the core of the parser
The previous parser suffered from a number of problems, most of which stemmed
from my own misunderstanding of what supplies can do combined with a lack of
knowledge about supply internals. I'm learning more and this should be a much
more sensible solution.

However, it is not yet completely working. The HTTP/1.0 tests without keep-alive
seem to be doing alright, but the HTTP/1.1 with keep-alive are not doing so
well so far.