Fixing bugs with chunked parsing + concurrency
This is a big deal. Also adds a :debug option to parse-http, which is helpful
for the maintainer but is a lot of data.

The algorithm for parsing chunked bodies is very naïve and will almost certainly
do very bad things, like hang the program forever, if given bad input at this
point.