Complete rewrite modeled after Cro
So, I was having weird concurrency issues with the way I had the
whenever blocks arranged. I decided to look at what Cro does and the
overall structure was more elegant and certainly easier to troubleshoot.
I did not copy the code from there, but rearranged the code I'd already
written on those terms. It now bears a marked resemblance to the code
found with Cro::HTTP::RequestParser and Cro::HTTP::RawBodyParser
compunits.

Cheers to jnthn and Edument!